Extracted functions into common file
